Support for Python3 and partial support of MultiNest v3.0 with INS
pyMultiNest is now compatible both with Python 2.6 or higher and
with Python3. With this update, you can run MultiNest v3.0 via
pyMultiNest with the new Importance Nested Sampling (INS) enabled.
At the moment, analysis and plotting of the data generated by
MultiNest v3.0 can be done with pyMultiNest only if INS is disabled.

On branch master, changes committed:
	modified:   .gitignore
	modified:   README.rst
	modified:   doc/conf.py
	modified:   multinest_bridge/Makefile
	modified:   multinest_bridge/cnest.c
	modified:   pyapemost/__init__.py
	modified:   pyapemost/analyse.py
	modified:   pyapemost/run.py
	modified:   pyapemost/watch.py
	modified:   pyapemost_demo.py
	modified:   pycuba/__init__.py
	modified:   pymultinest/__init__.py
	modified:   pymultinest/analyse.py
	modified:   pymultinest/plot.py
	modified:   pymultinest/run.py
	modified:   pymultinest/watch.py
	modified:   pymultinest_demo.py
	modified:   setup.py